requirements.txt
python-telegram-bot==20.7
pillow==10.0.1
pillow-heif==0.13.0
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
psutil==5.9.6